requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.4
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-jazzmin==3.0.1
django-modeltranslation==0.19.12
django-phonenumber-field==8.0.0
django-resized==1.0.3
django-redis==5.4.0
djangorestframework==3.15.2
drf-flex-fields==1.0.2
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
phonenumbers==8.13.52
pillow==11.0.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
